Terminado

+++ src/Controllers/PistasController.php
@@ -1,11 +1,14 @@
 <?php
 
-require("src/Gateways/PistasGateway.php");
+require_once("src/Gateways/PistasGateway.php");
+require_once("src/Gateways/ReservasGateway.php");
 
 class PistasController {
     private PistasGateway $gateway;
+    private ReservasGateway $reservaGateway;
     public function __construct(private Database $database)
     {
+        $this->reservaGateway = new ReservasGateway($database);
         $this->gateway = new PistasGateway($database);
     }
 
@@ -49,6 +52,12 @@ class PistasController {
                     );
                 break;
             case "DELETE":
+                $errors = $this->getDeleteErrors($pista, false);
+                if(!empty($errors)){
+                    http_response_code(422);
+                    echo json_encode(["errors"=>$errors]);
+                    break;
+                }
                 $rows = $this->gateway->delete($id);
                 echo json_encode(
                     [
@@ -97,18 +106,24 @@ class PistasController {
 
         if ($is_new) {
             if(!isset($data["nombre"]) || empty($data["nombre"])){
-                $errors[]="Se requiere nombre de la pista";
+                $errors[]="Se requiere nombre de la pista";   
+            }elseif (!is_string($data["nombre"])){
+                $errors[] = "La pista debe tener un nombre válido.";
             }
-            if(!isset($data["tipo"]) || empty($data["nombre"])){
+            if(!isset($data["tipo"]) || empty($data["tipo"])){
                 $errors[]="Se requiere tipo de pista";
+            }elseif (!is_string($data["tipo"])){
+                $errors[] = "El tipo de pista debe tener un nombre válido.";
             }
-            if(!isset($data["max_jugadores"]) || empty($data["nombre"])){
-                $errors[]="Se requiere cantidad de personas que juegan";
-            }elseif($data["max_jugadores"]<2){
+            if(isset($data["max_jugadores"]) || !empty($data["max_jugadores"])){
+                if (filter_var($data["max_jugadores"],FILTER_VALIDATE_INT)===false){
+                    $errors[] = "La cantidad de jugadores debe ser un número entero.";
+                }
+                if( $data["max_jugadores"]<2)
                 $errors[]="Para poder reservar una pista, necesitas al menos 2 personas.";
             }
-            if(!isset($data["disponible"]) || empty($data["disponible"])){
-                $errors[]="Se necesita saber si la pista está disponible";
+            if(isset($data["disponible"]) && !is_bool($data["disponible"])){
+                $errors[]="El campo Disponible debe se ser true o false";
             }
         } else {
             if(array_key_exists("nombre", $data)){
@@ -126,13 +141,23 @@ class PistasController {
                     $errors[] = "La cantidad de jugadores debe ser un número entero.";
                 }
             }
-            if(array_key_exists("nombre", $data)){
-                if (filter_var($data["disponible"],FILTER_VALIDATE_BOOL)===false){
+            if(array_key_exists("disponible", $data)){
+                if (!is_bool($data["disponible"])){
                     $errors[] = "La pista debe estar disponible (true) o no (false)";
                 }
             }
         }
 
+        return $errors;
+    }
+
+    private function getDeleteErrors(array $data){
+        $errors=[];
+        $reservaExistente = array_filter($this->reservaGateway->getAll(), fn($reserva) => $reserva["pista"]===$data["id"]);
+
+        if($reservaExistente){
+            $errors[] = "No se puede eliminar una pista si existen reservas a su nombre. Elimina las reservas primero.";
+        }
         return $errors;
     }
 }
